rename hooks to kebabcase
